requirements.txt
openai-whisper
torch
srt
deep-translator